treat primary keys as not-nullable
According to the SQL specification, primary keys must not be null.
However, xo treats primary keys as nullable unless "NOT NULL" appears
explicitly in the column schema. This can result in extra clunkiness
and/or illegal statements in generated code when it assumes a primary
key might be null.

Add a check for !IsPrimaryKey to all code locations where xo presently
checks for IsNullable.